+++ changelog.md
@@ -1,3 +1,22 @@
+## v4.2.0
+
++ Limit report_jobs to only display for current user is not an admin and authenticated
+
+## v4.1.0
+
++ Add `admin_only` column to reports, defaulted to `false`
++ Hide sensitive query fields when not authenticated as an admin
+
+## v4.0.1
+
++ Report Jobs retrieve the authenticatable model attribute name
+
+## v4.0.0
+
++ Add extended parameters to report_jobs queries (e.g. report_name)
++ Remove dated vue components
++ Add typescript definitions file
+
 ## v3.2.0
 
 + Add report run date to report parameters when export occurs